requirements.txt
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
colorlog==6.6.0
Cython==0.29.24
EasyProcess==0.3
entrypoint2==0.2.4
firebase-admin==5.1.0
google-api-core==2.2.2
google-api-python-client==2.32.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
grpcio-status==1.42.0
httplib2==0.20.2
idna==3.3
isodate==0.6.0
lxml==4.6.4
msgpack==1.0.3
mss==6.1.0
mutagen==1.45.1
numpy==1.21.4
opencv-python-headless==4.2.0.34
packaging==21.3
pafy==0.5.5
Pillow==8.4.0
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pycryptodome==3.12.0
pycryptodomex==3.12.0
pyparsing==3.0.6
pyscreenshot==3.0
PySocks==1.7.1
pyzmq==22.3.0
requests==2.26.0
rsa==4.8
simplejpeg==1.6.3
six==1.16.0
streamlink==3.0.3
tqdm==4.62.3
uritemplate==4.1.1
urllib3==1.26.7
vidgear==0.2.3
websocket-client==1.2.1
websockets==10.1
yt-dlp==2021.12.1
